fix: implemented login in index, redirect /home
