+++ client/styles/LoginStyled.js
@@ -8,11 +8,6 @@ export default styled.div`
   top: 0;
   bottom: 0;
   background-color: ${MODERN_GREY};
-  // background: linear-gradient(
-  //   90deg,
-  //   ${MODERN_PURPLE} 0%,
-  //   rgba(0, 255, 255) 75%
-  // );
 
   display: flex;
   align-items: center;
