+++ client/styles/index.js
@@ -1,9 +1,11 @@
+import AppStyled from "./AppStyles";
+import IndexStyled from "./IndexStyled";
 import LoadingStyled from "./LoadingStyled";
 import LoginStyled from "./LoginStyled";
 import ResultsStyled from "./ResultsStyled";
 import TeamStyled from "./TeamStyled";
 import LeagueStyled from "./LeagueStyled";
-import { IndexStyled, AppStyled } from "./AppStyles";
+import HomeStyled from "./HomeStyled";
 
 module.exports = {
   LoadingStyled,
@@ -11,6 +13,7 @@ module.exports = {
   ResultsStyled,
   TeamStyled,
   LeagueStyled,
-  IndexStyled,
   AppStyled,
+  HomeStyled,
+  IndexStyled,
 };
